Fix wmm-runtime path references and format-specifier warnings
Completes the rename-related fixes for src/wmm-runtime/ that didn't
make it into the previous rename commit, plus a set of pre-existing
bugs/warnings found and fixed while verifying the runtime rebuilds
cleanly post-rename.

- src/wmm-runtime/tests/{run_cas_semantics.sh,
  run_cas_unknown_feedback.sh}: ${REPO_ROOT}/Main/include updated to
  main/include (missed in the prior rename commit).

- src/wmm-runtime/{scheduler.c, wmm_hooks.c}: fixed hardcoded absolute
  #include "/home/aritra/fuzzer-1/FUZZER_Rebuilt/Main/include/shm_next_events.h"
  paths to .../SGF/main/include/shm_next_events.h. Still a
  machine-specific absolute path, not portable, but at least consistent
  with the new directory names; worth revisiting separately if this
  repo needs to build on a different machine/user.

- src/wmm-runtime/wmm_hooks.c: fixed a real bug at a pthread_join
  logging call -- printf(stderr, "...") was passing a FILE* as the
  format-string argument to printf (which takes no stream argument),
  rather than calling fprintf(stderr, "..."). Also fixed three
  %llx/%llu format-specifier mismatches against arguments cast to
  unsigned long instead of unsigned long long.

- src/wmm-runtime/eg.c: fixed one %lu/%llu mismatch in a node-lookup
  debug log, a stray literal "u" appended after a %d specifier
  (tid=%du, which just printed a literal 'u' character rather than
  being a real format flag), and several %d specifiers used against
  uint64_t (write_id/read_id) RF-edge logging arguments, corrected to
  %lu throughout.

- src/wmm-runtime/scheduler.c: fixed two remaining %d/%llu and
  %lld/%lu format-specifier mismatches in thread-registration and
  LOAD-event logging.

Verified: rm -f *.o libwmm_runtime.a; recompiled all 5 wmm-runtime
source files together (assert.c eg.c json.c scheduler.c wmm_hooks.c)
against the corrected include path -- zero warnings, down from the
~15 format-specifier warnings and the printf(stderr,...) bug present
before this commit. shm_next_events.cpp (which actually lives under
main/src/, not wmm-runtime/) compiled and archived into a fresh
libwmm_runtime.a successfully.

+++ SGF/src/wmm-runtime/scheduler.c
@@ -12,7 +12,7 @@
 #include <time.h>
 
 // TODO: This path is hardcoded - change it!
-#include "/home/aritra/fuzzer-1/FUZZER_Rebuilt/Main/include/shm_next_events.h"
+#include "/home/aritra/fuzzer-1/SGF/main/include/shm_next_events.h"
 
 /* eg.h keeps plain literals so it stays includable without going through
  * this hardcoded shm_next_events.h path; this is where the two vocabularies
@@ -1588,7 +1588,7 @@ void scheduler_thread_created(int tid,unsigned long long parentid) {
         pthread_cond_broadcast(&sched_cond);
     }
     add_to_all_threads(tid,parentid);
-    fprintf(stderr,"Registered thread=%d with parent=%d\n",tid,parentid);
+    fprintf(stderr,"Registered thread=%d with parent=%llu\n",tid,parentid);
     pthread_mutex_unlock(&sched_lock);
 }
 
@@ -1861,7 +1861,7 @@ bool scheduler_on_load_bytes_ex(void *addr, Access_Mode order,
                scheduler_terminate_locked(WMM_EXIT_INVALID_INPUT);
             }
             (void)mark_node_covered(node->id);
-            fprintf(stderr,"Executing the LOAD event, tid=%d,iid=%lld,vid=%d\n",tid,iid,vid);
+            fprintf(stderr,"Executing the LOAD event, tid=%llu,iid=%lld,vid=%d\n",(unsigned long long)tid,iid,vid);
             update_last_executed_event(tid,iid,vid);
         }
     }
